// frontend/src/api/generated/models/personalBestDto.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * TrainCoach API
 * OpenAPI spec version: v1
 */
import type { BestEffortType } from './bestEffortType';
import type { PersonalBestStepDto } from './personalBestStepDto';
import type { SportType } from './sportType';

export interface PersonalBestDto {
  sport?: SportType;
  type?: BestEffortType;
  value?: number;
  activityId?: string;
  /** @nullable */
  activityTitle?: string | null;
  achievedAtUtc?: string;
  isPrecise?: boolean;
  effortCount?: number;
  /** @nullable */
  progression?: PersonalBestStepDto[] | null;
}